Plans with verification

+++ src/database/migrations/20191203221151-create-students.js
@@ -21,11 +21,11 @@ module.exports = {
         allowNull: false,
       },
       height: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(10, 2),
         allowNull: false,
       },
       weight: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(10, 2),
         allowNull: false,
       },
       created_at: {
